\name{InfoTrad-package}
\alias{InfoTrad-package}
\alias{InfoTrad}
\docType{package}
\title{
\packageTitle{InfoTrad}
}
\description{
\packageDescription{InfoTrad}
}
%\details{
%The DESCRIPTION file:
%\packageDESCRIPTION{InfoTrad}
%\packageIndices{InfoTrad}
%}
\author{
\packageAuthor{InfoTrad}

Maintainer: \packageMaintainer{InfoTrad}
}
\references{
D. Easley, N. M. Kiefer, M. O'Hara, and J.B. Paperman. Liquidity, information, and infrequently traded stocks. The Journal of Finance, 51(4):1405-1436, 1996

D. Easley, S. Hvidkjaer, and M. O'Hara. Is information risk a determinant of asset returns? The Journal of Finance, 57(5): 2185-2221, 2002. 

D. Easley, S. Hvidkjaer, and M. O'Hara. Factoring information into returns. Journal of Financial and Quantitative Analysis, 45(2): 293-309, 2010. 

Ersan, O., and Alici, A. An unbiased computation methodology for estimating the probability of informed trading (PIN). Journal of International Financial Markets, Institutions and Money, 43(1):74-94, 2016.

Q. Gan, W.C. Wei, and D. Johnstone. A faster estimation method for the probability of informed trading using hierarhical agglomerative clustering. Quantitative Finance, 15(11):1805-1821, 2015

H.W. W. Lin and W. C. Ke. A computing bias in estimating the probability of informed trading. Journal of Financial Markets, 14(4):625-640, 2011.

D. Mullner. fastcluster: Fast hierarchical, agglomerative clustering routines for r and python. Journal of Statistical Software, 53(9):1-18, 2013.

Y. Yan and S. Zhang. An improved estimation method and empirical properties of the probability of informed trading. Journal of Banking & Finance, 36(2):454-467, 2012.
}
\keyword{Probability of Informed Trading (PIN), LK factorization, EHO factorization, YZ Algorithm, CPIN}

